streaming

+++ mashupmedia-server/src/main/java/org/mashupmedia/dto/media/playlist/NavigatePlaylistPayload.java
@@ -11,5 +11,5 @@ import lombok.NoArgsConstructor;
 @Builder(toBuilder = true)
 public class NavigatePlaylistPayload {
     private NavigatePlaylistType navigatePlaylistType;
-    private Integer mediaItemId;
+    private Long mediaItemId;
 }
